Add possibility to personalize capabilityString in function of (browserName, platform, version).
I had some problem due to the structure of the 'capabilityString'.
Chrome is updated very often, and consequently at each update there is a missmatch
between the screenshot folder and the 'capabilityString' due to the version value.

As first solution, we try to block the update of Chrome using all the various tutorials
that we can find on the web, unfortunately any of these solutions worked.

Consequently, I add a new optional parameter called 'screenshotCapabilityFn' which is
a function that takes as argument (browerName, platform, version) and return the value
for the 'capabilityString'. By default, this function return the same string like before.
I use 'screenshotCapabilityFn' in order to keep only the first number of a version,
and thus the missmatch problem occurs only with big version changes which is perfectly
fine for me.